+++ CHANGELOG.md
@@ -7,9 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Tests for more message types.
+
 ### Fixed
 
-- Parsing of *Tune Request* message
+- Parsing of *Tune Request* message.
 
 ## [0.1.0] - 2022-12-18